Moves EbenefitsBnftClaimStatusWebService to it's own file. Updates v1 app controller and claim spec.
