+++ modules/claims_api/app/clients/claims_api/bgs_client/definitions.rb
@@ -64,7 +64,7 @@ module ClaimsApi
       ##
       # EBenefitsBnftClaimStatusWebServiceBean
       #
-      module EBenefitsBenefitClaimStatusWebServiceBean
+      module EBenefitsBnftClaimStatusWebService
         DEFINITION =
           Bean.new(
             path: 'EBenefitsBnftClaimStatusWebServiceBean',
@@ -73,23 +73,6 @@ module ClaimsApi
               data: nil
             )
           )
-      end
-
-      module EBenefitsBenefitClaimStatusWebService
-        DEFINITION =
-          Service.new(
-            bean: EBenefitsBenefitClaimStatusWebServiceBean::DEFINITION,
-            path: 'EBenefitsBnftClaimStatusWebService'
-          )
-
-        module FindBenefitClaimsStatusByParticipantId
-          DEFINITION =
-            Action.new(
-              service: EBenefitsBenefitClaimStatusWebService::DEFINITION,
-              name: 'findBenefitClaimsStatusByPtcpntId',
-              key: 'BenefitClaimsDTO'
-            )
-        end
       end
 
       ##
