+++ modules/claims_api/spec/requests/v1/claims_spec.rb
@@ -3,6 +3,7 @@
 require 'rails_helper'
 require_relative '../../rails_helper'
 require 'bgs/power_of_attorney_verifier'
+require 'bgs_service/e_benefits_bnft_claim_status_web_service'
 
 RSpec.describe 'ClaimsApi::V1::Claims', type: :request do
   include SchemaMatchers
@@ -35,7 +36,7 @@ RSpec.describe 'ClaimsApi::V1::Claims', type: :request do
   end
   let(:claims_service) do
     if Flipper.enabled? :claims_status_v1_bgs_enabled
-      ClaimsApi::LocalBGS
+      ClaimsApi::EbenefitsBnftClaimStatusWebService
     else
       ClaimsApi::UnsynchronizedEVSSClaimService
     end
